requirements.txt
certifi
click
Flask
Flask-SQLAlchemy
gunicorn
deepface
tensorflow
Jinja2
numpy
Werkzeug
mysql-connector-python